// src/app/models/employee.model.ts
export class Employee {
    id: number;
    name: string;
    gender: string;
    contactPreference: string;
    email?: string;
    phoneNumber?: number;
    dateOfBirth: Date;
    department: string;
    department1?: string;
    isActive: boolean;
    photoPath?: string;
    password?: string;
    confirmPassword?: string;

}